requirements.txt
asgiref==3.2.3
Django==3.0.1
django-crispy-forms==1.9.0
django-six==1.0.4
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
